requirements.txt
spacy>=3.4.0
pandas>=1.3.0
nltk>=3.7
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
safetensors>=0.3.0
sentencepiece>=0.1.99
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.3.0
fastapi>=0.100.0
python-multipart>=0.0.20
